VISoR_Reconstruction/requirements.txt
h5py
numpy
opencv-python
pillow
pyqt5
pyyaml
scipy
scikit-learn
SimpleITK-Elastix
tifffile
torch
torchvision

